requirements.txt
runpod>=1.7.0
huggingface_hub>=0.23.0
hf_transfer>=0.1.6
llama-cpp-python>=0.3.0
pydantic>=2.0
pydantic-settings>=2.0
uvicorn>=0.29.0
